pyproject.toml
[project]
name = "sucre"
version = "0.1.0"
description = "A pipeline for training glucometric predictive models"
authors = [
    {name = "James Quang"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "pycaret (>=3.3.2,<4.0.0)",
    "typer (>=0.20.0,<0.21.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pytest = "^8.4.2"


[tool.poetry.scripts]
sucre = "sucre.main:app"

